♻️ Refactored progress dragging play state into hook
