+++ packages/custoplayer/src/lib/components/Duration.tsx
@@ -5,7 +5,6 @@ import { formattedDurationAtom, myScope } from '../atoms';
 function Duration() {
   const formattedDuration = useAtomValue(formattedDurationAtom, myScope);
   // When the length is < 60 minutes the length is 5 otherwise it is 7
-  console.log(formattedDuration.length);
   return (
     <TimeText
       isLargerThan60Minutes={formattedDuration.length === 7}
